refactor: move record scopes to a single place
